refactor(nutrition): extract catalog persistence helpers
